Edit : annotated test
